user authorization for edit and delete

+++ mean-course/backend/routes/posts.js
@@ -33,7 +33,8 @@ const url = req.protocol + '://' + req.get("host");
   const post = new Post({
     title: req.body.title,
     content: req.body.content,
-    imagePath: url + "/images/" + req.file.filename
+    imagePath: url + "/images/" + req.file.filename,
+    creator: req.userData.userId
   });
   post.save().then(createdPost => {
     res.status(201).json({
@@ -58,9 +59,12 @@ router.put("/:id", checkAuth, multer({storage: storage}).single("image"), (req, 
     content: req.body.content,
     imagePath: imagePath
   });
-  console.log(post);
-  Post.updateOne({ _id: req.params.id }, post).then(result => {
-    res.status(200).json({ message: "Update successful!" });
+  Post.updateOne({ _id: req.params.id, creator: req.userData.userId }, post).then(result => {
+    if (result.nModified > 0) {
+      res.status(200).json({ message: "Update successful!" });
+    } else {
+    res.status(401).json({ message: "Not Authorized!" });
+  }
   });
 });
 
@@ -98,10 +102,14 @@ router.get("/:id", (req, res, next) => {
 });
 
 router.delete("/:id", checkAuth, (req, res, next) => {
-  Post.deleteOne({ _id: req.params.id }).then(result => {
-    console.log(result);
-    res.status(200).json({ message: "Post deleted!" });
-  });
+  Post.deleteOne({ _id: req.params.id, creator: req.userData.userId }).then(result => {
+    if (result.n > 0) {
+      res.status(200).json({ message: "Deletion successful!" });
+    } else {
+    res.status(401).json({ message: "Not Authorized!" });
+    }
+  }
+);
 });
 
 module.exports = router;
